iOS: Bump version to 5.1.2
This commit bumps the `ios` branch from 5.0.0 alpha to 5.1.2 release.

I merged v5.0-release first then v5.1-release. So it was a sequential
merge.

I'm just waiting to the approval of continuing development.

The stuff that are missing (those are so close to implement):
- #145484 iOS: Keyboard Input Support: https://projects.blender.org/blender/blender/pulls/145484
- #145702 iOS: Add clipboard Image: https://projects.blender.org/blender/blender/pulls/145702
- Fix QOS threading
- Expose Downloads
- UIScene Migration

`make deps` needs to run two times. (still hacky....) OSL occurs racing
conditions.

Disclaimer: I used AI for searching in Documentation of Deps. So the
implementation is by hand. **SACRED PINKY FINGER PROMISE**

Pull Request: https://projects.blender.org/blender/blender/pulls/161073

+++ source/blender/editors/object/interface_template_shape_key_tree.cc
@@ -23,6 +23,7 @@
 
 #include "DEG_depsgraph.hh"
 
+#include "DEG_depsgraph_build.hh"
 #include "DNA_key_types.h"
 
 #include "WM_api.hh"
@@ -69,13 +70,34 @@ class ShapeKeyDragController : public ui::AbstractViewItemDragController {
 
   void *create_drag_data() const override
   {
-    ShapeKey *drag_data = MEM_callocN<ShapeKey>(__func__);
-    *drag_data = drag_key_;
-    return drag_data;
-  }
-  void on_drag_start(bContext & /*C*/) override
-  {
-    drag_key_.object->shapenr = drag_key_.index + 1;
+    int selected_count = [&]() -> int {
+      int count = 0;
+      for (KeyBlock &kb : drag_key_.key->block) {
+        count += (kb.flag & KEYBLOCK_SEL) != 0;
+      }
+      return count;
+    }();
+
+    /* Allocate one extra element, to use it as null-delimiter. */
+    KeyBlock **selected_keys_ = MEM_new_array_zeroed<KeyBlock *>(selected_count + 1,
+                                                                 "Selected Key Blocks");
+
+    selected_count = 0;
+
+    for (const auto [index, kb] : drag_key_.key->block.enumerate()) {
+      if (index == 0) {
+        /* Prevent basis shape key from dragging. */
+        continue;
+      }
+
+      if (kb.flag & KEYBLOCK_SEL) {
+        selected_keys_[selected_count] = &kb;
+        selected_count++;
+      }
+    }
+    BLI_assert_msg(selected_keys_[selected_count] == nullptr,
+                   "Expected last element to be null (null-delimiter)");
+    return selected_keys_;
   }
 };
 
@@ -98,17 +120,18 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
     if (drag.type != WM_DRAG_SHAPE_KEY) {
       return false;
     }
-    const ShapeKey *drag_shapekey = static_cast<const ShapeKey *>(drag.poin);
-    if (drag_shapekey->index == drop_index_) {
+
+    const KeyBlock **drag_shapekey = static_cast<const KeyBlock **>(drag.poin);
+    if (!drag_shapekey || !drag_shapekey[0]) {
       return false;
     }
+
     return true;
   }
 
   std::string drop_tooltip(const ui::DragInfo &drag_info) const override
   {
-    const ShapeKey *drag_shapekey = static_cast<const ShapeKey *>(drag_info.drag_data.poin);
-    const StringRef drag_name = drag_shapekey->kb->name;
+    const StringRef drag_name = TIP_("Selected Keys");
     const StringRef drop_name = drop_kb_.name;
 
     switch (drag_info.drop_location) {
@@ -132,10 +155,12 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
 
   bool on_drop(bContext *C, const ui::DragInfo &drag_info) const override
   {
-    const ShapeKey *drag_shapekey = static_cast<const ShapeKey *>(drag_info.drag_data.poin);
-    int drop_index = drop_index_;
-    const int drag_index = drag_shapekey->index;
+    Object *ob = CTX_data_active_object(C);
+    Key *key = BKE_key_from_object(ob);
+    const KeyBlock **drag_shapekey = static_cast<const KeyBlock **>(drag_info.drag_data.poin);
 
+    const int first_drag_index = BLI_findindex(&key->block, drag_shapekey[0]);
+    int drop_index = BLI_findindex(&key->block, &drop_kb_);
     switch (drag_info.drop_location) {
       case ui::DropLocation::Into:
         BLI_assert_unreachable();
@@ -144,17 +169,27 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
         if (drop_index == 0) {
           return false;
         }
-        drop_index -= int(drag_index < drop_index);
+        drop_index -= int(first_drag_index < drop_index);
         break;
       case ui::DropLocation::After:
-        drop_index += int(drag_index > drop_index);
+        drop_index += int(first_drag_index > drop_index);
         break;
     }
-    Object *object = drag_shapekey->object;
-    BKE_keyblock_move(object, drag_shapekey->index, drop_index);
 
-    DEG_id_tag_update(static_cast<ID *>(object->data), ID_RECALC_GEOMETRY);
-    WM_event_add_notifier(C, NC_OBJECT | ND_DRAW, object);
+    for (int8_t i = 0; drag_shapekey[i] != nullptr; i++) {
+      const int drag_index = BLI_findindex(&key->block, drag_shapekey[i]);
+      if (drag_index == -1) {
+        continue;
+      }
+      if (i > 0) {
+        /* Place subsequent items directly after the previously moved item. */
+        drop_index += int(drag_index > drop_index);
+      }
+      BKE_keyblock_move(ob, drag_index, drop_index);
+    }
+
+    DEG_id_tag_update(ob->data, ID_RECALC_GEOMETRY);
+    WM_event_add_notifier(C, NC_OBJECT | ND_DRAW, ob);
     ED_undo_push(C, "Drop Active Shape Key");
 
     return true;
@@ -175,20 +210,28 @@ class ShapeKeyItem : public ui::AbstractTreeViewItem {
     shape_key_.index = index;
   };
 
-  void build_row(uiLayout &row) override
+  void build_row(ui::Layout &row) override
   {
     uiItemL_ex(&row, this->label_, ICON_SHAPEKEY_DATA, false, false);
-    uiLayout *sub = &row.row(true);
-    sub->use_property_decorate_set(false);
+    ui::Layout &sub = row.row(true);
+    sub.use_property_decorate_set(false);
     PointerRNA shapekey_ptr = RNA_pointer_create_discrete(
-        &shape_key_.key->id, &RNA_ShapeKey, shape_key_.kb);
+        &shape_key_.key->id, RNA_ShapeKey, shape_key_.kb);
 
-    if (shape_key_.index > 0) {
-      sub->prop(&shapekey_ptr, "value", UI_ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+    if (shape_key_.key->type == KEY_NORMAL) {
+      sub.prop(&shapekey_ptr, "frame", ui::ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+    }
+    else {
+      if (shape_key_.index > 0) {
+        sub.prop(&shapekey_ptr, "value", ui::ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+      }
     }
 
-    sub->prop(&shapekey_ptr, "mute", UI_ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
-    sub->prop(&shapekey_ptr, "lock_shape", UI_ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+    sub.prop(&shapekey_ptr, "mute", ui::ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+    sub.prop(&shapekey_ptr, "lock_shape", ui::ITEM_R_ICON_ONLY, std::nullopt, ICON_NONE);
+    if (shape_key_.kb->flag & KEYBLOCK_MUTE) {
+      row.active_set(false);
+    }
   }
 
   std::optional<bool> should_be_active() const override
@@ -199,7 +242,7 @@ class ShapeKeyItem : public ui::AbstractTreeViewItem {
   void on_activate(bContext &C) override
   {
     PointerRNA object_ptr = RNA_pointer_create_discrete(
-        &shape_key_.object->id, &RNA_Object, shape_key_.object);
+        &shape_key_.object->id, RNA_Object, shape_key_.object);
     PropertyRNA *prop = RNA_struct_find_property(&object_ptr, "active_shape_key_index");
     RNA_property_int_set(&object_ptr, prop, shape_key_.index);
     RNA_property_update(&C, &object_ptr, prop);
@@ -226,7 +269,7 @@ class ShapeKeyItem : public ui::AbstractTreeViewItem {
   bool rename(const bContext &C, StringRefNull new_name) override
   {
     PointerRNA shapekey_ptr = RNA_pointer_create_discrete(
-        &shape_key_.key->id, &RNA_ShapeKey, shape_key_.kb);
+        &shape_key_.key->id, RNA_ShapeKey, shape_key_.kb);
     RNA_string_set(&shapekey_ptr, "name", new_name.c_str());
     ED_undo_push(const_cast<bContext *>(&C), "Rename shape key");
     return true;
@@ -242,25 +285,22 @@ class ShapeKeyItem : public ui::AbstractTreeViewItem {
     Main *bmain = CTX_data_main(C);
     BKE_object_shapekey_remove(bmain, shape_key_.object, shape_key_.kb);
     DEG_id_tag_update(&shape_key_.object->id, ID_RECALC_GEOMETRY);
+    DEG_relations_tag_update(CTX_data_main(C));
     WM_event_add_notifier(C, NC_OBJECT | ND_DRAW, nullptr);
     ED_undo_grouped_push(C, "Delete Shape Key");
   }
 
-  void build_context_menu(bContext &C, uiLayout &layout) const override
+  void build_context_menu(bContext &C, ui::Layout &layout) const override
   {
     MenuType *mt = WM_menutype_find("MESH_MT_shape_key_tree_context_menu", true);
     if (!mt) {
       return;
     }
-    UI_menutype_draw(&C, mt, &layout);
+    ui::menutype_draw(&C, mt, &layout);
   }
 
   std::unique_ptr<ui::AbstractViewItemDragController> create_drag_controller() const override
   {
-    if (shape_key_.index == 0) {
-      /* Prevent basis shape key from dragging. */
-      return nullptr;
-    }
     return std::make_unique<ShapeKeyDragController>(
         static_cast<ShapeKeyTreeView &>(get_tree_view()), shape_key_);
   }
@@ -278,22 +318,22 @@ void ShapeKeyTreeView::build_tree()
   if (key == nullptr) {
     return;
   }
-  int index = 1;
-  LISTBASE_FOREACH_INDEX (KeyBlock *, kb, &key->block, index) {
-    this->add_tree_item<ShapeKeyItem>(&object_, key, kb, index);
+
+  for (const auto [index, kb] : key->block.enumerate()) {
+    this->add_tree_item<ShapeKeyItem>(&object_, key, &kb, index);
   }
 }
 
-void template_tree(uiLayout *layout, bContext *C)
+void template_tree(ui::Layout *layout, bContext *C)
 {
   Object *ob = CTX_data_active_object(C);
   if (ob == nullptr) {
     return;
   }
 
-  uiBlock *block = layout->block();
+  ui::Block *block = layout->block();
 
-  ui::AbstractTreeView *tree_view = UI_block_add_view(
+  ui::AbstractTreeView *tree_view = block_add_view(
       *block,
       "Shape Key Tree View",
       std::make_unique<ed::object::shapekey::ShapeKeyTreeView>(*ob));
